move usePolling to hooks directory
